rename the "http" module to "app"
