+++ test/integration/test_http_api.py
@@ -10,7 +10,7 @@ configuration.ratelimit["read"]["capacity"] = 2**64 - 1
 configuration.ratelimit["create"]["capacity"] = 2**64 - 1
 configuration.ratelimit["delete"]["capacity"] = 2**64 - 1
 
-from pinnwand import configuration, database, http, utility
+from pinnwand import configuration, database, app, utility
 
 
 class DeprecatedAPITestCase(tornado.testing.AsyncHTTPTestCase):
@@ -19,7 +19,7 @@ class DeprecatedAPITestCase(tornado.testing.AsyncHTTPTestCase):
         database.Base.metadata.create_all(database._engine)
 
     def get_app(self) -> tornado.web.Application:
-        return http.make_application()
+        return app.make_application()
 
     def test_api_new(self) -> None:
         response = self.fetch(
@@ -291,7 +291,7 @@ class APIv1TestCase(tornado.testing.AsyncHTTPTestCase):
         database.Base.metadata.create_all(database._engine)
 
     def get_app(self) -> tornado.web.Application:
-        return http.make_application()
+        return app.make_application()
 
     def test_api_new(self) -> None:
         response = self.fetch(
