+++ test/integration/test_http_curl.py
@@ -10,7 +10,7 @@ configuration.ratelimit["read"]["capacity"] = 2**64 - 1
 configuration.ratelimit["create"]["capacity"] = 2**64 - 1
 configuration.ratelimit["delete"]["capacity"] = 2**64 - 1
 
-from pinnwand import database, http
+from pinnwand import database, app
 
 
 class CurlTestCase(tornado.testing.AsyncHTTPTestCase):
@@ -19,7 +19,7 @@ class CurlTestCase(tornado.testing.AsyncHTTPTestCase):
         database.Base.metadata.create_all(database._engine)
 
     def get_app(self) -> tornado.web.Application:
-        return http.make_application()
+        return app.make_application()
 
     def test_curl_post_no_lexer(self) -> None:
         response = self.fetch(
